requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
requests==2.32.0
psycopg2-binary==2.9.9
python-docx==1.1.2
python-multipart==0.0.9
pydantic==2.7.0
python-jose==3.3.0
passlib==1.7.4
aiofiles==23.2.1
